Differentiate MRI and PET + sanctuarize code for MICCAI submission
