+++ deformetrica/LAE_experiments/ADNI_data/model_MRI.xml
@@ -7,13 +7,13 @@
 		<exponential-type>euclidean</exponential-type>
 		<number-of-sources>15</number-of-sources>
 		<t0>74.4</t0>
-		<v0>output_CVAE_16_MRI_15_big/LongitudinalMetricModel_v0.txt</v0>
-		<p0>output_CVAE_16_MRI_15_big/LongitudinalMetricModel_p0.txt</p0>
+		<v0>output_CVAE_16_MRI_15_sources/LongitudinalMetricModel_v0.txt</v0>
+		<p0>output_CVAE_16_MRI_15_sources/LongitudinalMetricModel_p0.txt</p0>
 	</deformation-parameters>
-	<initial-sources>output_CVAE_16_MRI_15_big/LongitudinalMetricModel_sources.txt</initial-sources>
-	<initial-modulation-matrix>output_CVAE_16_MRI_15_big/LongitudinalMetricModel_modulation_matrix.txt</initial-modulation-matrix>
-	<initial-onset-ages>output_CVAE_16_MRI_15_big/LongitudinalMetricModel_onset_ages.txt</initial-onset-ages>
-	<initial-log-accelerations>output_CVAE_16_MRI_15_big/LongitudinalMetricModel_log_accelerations.txt</initial-log-accelerations>
+	<initial-sources>output_CVAE_16_MRI_15_sources/LongitudinalMetricModel_sources.txt</initial-sources>
+	<initial-modulation-matrix>output_CVAE_16_MRI_15_sources/LongitudinalMetricModel_modulation_matrix.txt</initial-modulation-matrix>
+	<initial-onset-ages>output_CVAE_16_MRI_15_sources/LongitudinalMetricModel_onset_ages.txt</initial-onset-ages>
+	<initial-log-accelerations>output_CVAE_16_MRI_15_sources/LongitudinalMetricModel_log_accelerations.txt</initial-log-accelerations>
 </data-set>
